[IMP][partner_perception][7614]
Migration 13 - 11
