+++ partner_perception/models/__init__.py
@@ -1,1 +1,1 @@
-from . import account_move
+from . import account_invoice
