+++ partner_perception/views/account_move_view.xml
@@ -1,11 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
 <odoo>
 	<record id="partner_perception_view_move_form_inherit" model="ir.ui.view">
-	    <field name="model">account.move</field>
+	    <field name="model">account.invoice</field>
 	    <field name="name">Account Invoice</field>
-	    <field name="inherit_id" ref="account.view_move_form"/>
+	    <field name="inherit_id" ref="account.invoice_form"/>
 	    <field name="arch" type="xml">
-	        <xpath expr="//button[@name='preview_invoice']" position="after">
+	        <xpath expr="//button[@name='action_invoice_open']" position="after">
 				<button name="partner_perceptions" type="object" string="Partner Perception" />
 	        </xpath>
 	    </field>
